requirements.txt
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.8
gcloud==0.17.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
imageio==2.9.0
imageio-ffmpeg==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
moviepy==1.0.3
numpy==1.19.1
oauth2client==3.0.0
Pillow==7.2.0
proglog==0.1.9
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
tqdm==4.48.2
Werkzeug==1.0.1
